Several things
After the change we made yesterday, it wouldn't even compile. As far as
I understand, both structs (Board and PossibleMove) must be defined in
the same header. Otherwise, it's a never-ending circular definition
(Board.h uses PossibleMove which is defined in PossibleMove.h, which
uses Board which is defined in Board.h, etc. ).

Also:
-added a little safeguard against illegal input in readTile.
-tiny change in getScore that makes the minimax run much faster