Two small but meaningful fixes
The change in board.c prevents situations where the king is supposedly
in check, but actually the path between it the and the piece threatening
it is blocked by other enemy pieces.

The change in chess.c fixes a problem with loading saved games where the
game mode was 2 (Human vs. AI) - it wouldn't load the board and the rook
and king movement tracking arrays.